refactor: move tests for external remove proposals
We want to keep them, but in the regular `conversation::proposal` module
