+++ crypto/src/transaction_context/conversation/external_proposal.rs
@@ -36,225 +36,3 @@ impl TransactionContext {
             .map_err(Into::into)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::test_utils::*;
-
-    mod add {
-        use super::*;
-
-        #[apply(all_cred_cipher)]
-        async fn guest_should_externally_propose_adding_itself_to_owner_group(case: TestContext) {
-            let [owner, guest] = case.sessions().await;
-            Box::pin(async move {
-                let conversation = case.create_conversation([&owner]).await;
-
-                // Craft an external proposal from guest
-                // Owner receives external proposal message from server
-                let (proposal_guard, decrypted) = conversation
-                    .external_join_proposal(&guest)
-                    .await
-                    .notify_member_fallible(&owner)
-                    .await;
-                let conversation = proposal_guard.finish();
-                // just owner for now
-                assert_eq!(conversation.member_count().await, 1);
-
-                // verify Guest's (sender) identity
-                guest
-                    .verify_sender_identity(&case, &guest.initial_credential, &decrypted.unwrap())
-                    .await;
-
-                // simulate commit message reception from server
-                let conversation = conversation.commit_pending_proposals_notify().await;
-                // guest joined the group
-                assert_eq!(conversation.member_count().await, 2);
-
-                // guest can send messages in the group
-                assert!(conversation.is_functional_and_contains([&owner, &guest]).await);
-            })
-            .await
-        }
-    }
-
-    mod remove {
-        use openmls::prelude::{ProcessMessageError, ValidationError};
-
-        use super::*;
-        use crate::{MlsError, MlsErrorKind};
-
-        #[apply(all_cred_cipher)]
-        async fn ds_should_remove_guest_from_conversation(mut case: TestContext) {
-            let [owner, guest, ds] = case.sessions().await;
-            Box::pin(async move {
-                let conversation = case
-                    .create_conversation_with_external_sender(&ds, [&owner, &guest])
-                    .await;
-                assert_eq!(conversation.member_count().await, 2);
-
-                // now, as e.g. a Delivery Service, let's create an external remove proposal
-                // and kick guest out of the conversation
-
-                let conversation = conversation.external_remove_proposal_notify(&ds, &guest).await;
-                let conversation = conversation.commit_pending_proposals_notify().await;
-                assert_eq!(conversation.member_count().await, 1);
-
-                // guest can no longer participate
-                assert!(guest.transaction.conversation(conversation.id()).await.is_err());
-                assert!(!conversation.can_talk(&owner, &guest).await);
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        async fn should_fail_when_invalid_external_sender(mut case: TestContext) {
-            use crate::mls;
-
-            let [owner, guest, ds, attacker] = case.sessions().await;
-            Box::pin(async move {
-                // Delivery service key is used in the group..
-                let conversation = case
-                    .create_conversation_with_external_sender(&ds, [&owner, &guest])
-                    .await;
-                assert_eq!(conversation.member_count().await, 2);
-
-                // now, attacker will try to remove guest from the group, and should fail
-                let proposal_guard = conversation
-                    .external_remove_proposal_with_sender_index(&attacker, 1, &guest)
-                    .await;
-
-                let (proposal_guard, owner_decrypt) = proposal_guard.notify_member_fallible(&owner).await;
-
-                assert!(matches!(
-                    owner_decrypt.unwrap_err(),
-                    mls::conversation::Error::Mls(MlsError {
-                        source: MlsErrorKind::MlsMessageError(ProcessMessageError::ValidationError(
-                            ValidationError::UnauthorizedExternalSender
-                        )),
-                        ..
-                    })
-                ));
-
-                let (_, guest_decrypt) = proposal_guard.notify_member_fallible(&guest).await;
-                assert!(matches!(
-                    guest_decrypt.unwrap_err(),
-                    mls::conversation::Error::Mls(MlsError {
-                        source: MlsErrorKind::MlsMessageError(ProcessMessageError::ValidationError(
-                            ValidationError::UnauthorizedExternalSender
-                        )),
-                        ..
-                    })
-                ));
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        async fn should_fail_when_wrong_signature_key(mut case: TestContext) {
-            use crate::mls;
-
-            let [owner, guest, ds] = case.sessions().await;
-            Box::pin(async move {
-                // Here we're going to add the Delivery Service's (DS) signature key to the
-                // external senders list. However, for the purpose of this test, we will
-                // intentionally _not_ use that key when generating the remove proposal below.
-                let conversation = case
-                    .create_conversation_with_external_sender(&ds, [&owner, &guest])
-                    .await;
-                assert_eq!(conversation.member_count().await, 2);
-
-                // Intentionally use the guest's credential, and therefore the guest's signature
-                // key when generating the proposal so that the signature verification fails.
-                let proposal_guard = conversation.external_remove_proposal(&guest, &guest).await;
-
-                let (proposal_guard, owner_decrypt) = proposal_guard.notify_member_fallible(&owner).await;
-                assert!(matches!(
-                    owner_decrypt.unwrap_err(),
-                    mls::conversation::Error::Mls(MlsError {
-                        source: MlsErrorKind::MlsMessageError(ProcessMessageError::InvalidSignature),
-                        ..
-                    })
-                ));
-
-                let (_, guest_decrypt) = proposal_guard.notify_member_fallible(&guest).await;
-                assert!(matches!(
-                    guest_decrypt.unwrap_err(),
-                    mls::conversation::Error::Mls(MlsError {
-                        source: MlsErrorKind::MlsMessageError(ProcessMessageError::InvalidSignature),
-                        ..
-                    })
-                ));
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        async fn joiners_from_welcome_can_accept_external_remove_proposals(mut case: TestContext) {
-            let [alice, bob, charlie, ds] = case.sessions().await;
-            Box::pin(async move {
-                let conversation = case.create_conversation_with_external_sender(&ds, [&alice, &bob]).await;
-                assert_eq!(conversation.member_count().await, 2);
-
-                // Charlie joins through a Welcome and should get external_senders from Welcome
-                // message and not from configuration
-                // charlie can only get it from there, because the `MlsCustomgConfiguration` that they receive when
-                // processing the welcome, doesn't contain any info about an external sender.
-                let conversation = conversation.invite_notify([&charlie]).await;
-                assert_eq!(conversation.member_count().await, 3);
-                assert!(conversation.is_functional_and_contains([&alice, &bob, &charlie]).await);
-
-                // now, as e.g. a Delivery Service, let's create an external remove proposal
-                // and kick Bob out of the conversation
-
-                // joiner from Welcome should be able to verify the external remove proposal since
-                // it has fetched back the external_sender from Welcome
-                let conversation = conversation.external_remove_proposal_notify(&ds, &bob).await;
-                let conversation = conversation
-                    .acting_as(&charlie)
-                    .await
-                    .commit_pending_proposals_notify()
-                    .await;
-
-                assert_eq!(conversation.member_count().await, 2);
-
-                assert!(conversation.is_functional_and_contains([&alice, &charlie]).await);
-                assert!(!conversation.can_talk(&alice, &bob).await);
-            })
-            .await
-        }
-
-        #[apply(all_cred_cipher)]
-        async fn joiners_from_external_commit_can_accept_external_remove_proposals(mut case: TestContext) {
-            let [alice, bob, charlie, ds] = case.sessions().await;
-            Box::pin(async move {
-                let conversation = case.create_conversation_with_external_sender(&ds, [&alice, &bob]).await;
-                assert_eq!(conversation.member_count().await, 2);
-
-                // Charlie joins through an external commit and should get external_senders
-                // from the public group state and not from configuration
-                let conversation = conversation.external_join_notify(&charlie).await;
-                assert_eq!(conversation.member_count().await, 3);
-                assert!(conversation.is_functional_and_contains([&alice, &bob, &charlie]).await);
-
-                // now, as e.g. a Delivery Service, let's create an external remove proposal
-                // and kick Bob out of the conversation
-
-                // joiner from external commit should be able to verify the external remove proposal
-                // since it has fetched back the external_sender from external commit
-                let conversation = conversation.external_remove_proposal_notify(&ds, &bob).await;
-                let conversation = conversation
-                    .acting_as(&charlie)
-                    .await
-                    .commit_pending_proposals_notify()
-                    .await;
-
-                assert_eq!(conversation.member_count().await, 2);
-
-                assert!(conversation.is_functional_and_contains([&alice, &charlie]).await);
-                assert!(!conversation.can_talk(&alice, &bob).await);
-            })
-            .await
-        }
-    }
-}
